requirements.txt
pandas>=1.3.0
numpy>=1.20.0
openai>=1.0.0
langchain>=0.0.267
langchain-openai>=0.0.1
langchain-community>=0.0.1
langchain-text-splitters>=0.0.1
scikit-learn>=1.0.0
tqdm>=4.60.0
rank_bm25>=0.2.2
faiss-cpu>=1.7.0
python-dotenv>=1.0.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
openpyxl>=3.0.0 